+++ src/app/options/options.component.ts
@@ -8,6 +8,7 @@ import {
     ViewChildren,
 } from '@angular/core';
 import {
+    AsyncPipe,
     JsonPipe,
     NgClass,
     NgFor,
@@ -24,17 +25,17 @@ import { IsEmptyPipe } from '@open-source/is-empty/is-empty.pipe';
 import { ColorPipe } from '@open-source/color/color.pipe';
 import { CardContentComponent } from '../shared/card-content/card-content.component';
 import { TrackByService } from '@libraries/track-by/track-by.service';
-import { ScrollService } from '@open-source/service/scroll.service';
-import { Router } from '@angular/router';
-import { debounceTime, fromEvent, takeUntil } from 'rxjs';
-import { UnSubscriber } from '@libraries/unsubscriber/unsubscriber.service';
+import { Observable } from 'rxjs';
+import { ScrollService } from '@open-source/scroll/scroll.service';
+import { OpenSourcePath } from '@open-source/path/open-source.path';
+import { ChangeAccordionService } from '../shared/service/change-accordion.service';
 
 @Component({
     selector: 'jsdaddy-open-source-options',
     templateUrl: './options.component.html',
     styleUrls: ['./options.component.scss'],
     standalone: true,
-    providers: [ScrollService],
+    providers: [ScrollService, ChangeAccordionService],
     imports: [
         JsonPipe,
         NgFor,
@@ -51,45 +52,25 @@ import { UnSubscriber } from '@libraries/unsubscriber/unsubscriber.service';
         IsEmptyPipe,
         ColorPipe,
         CardContentComponent,
+        AsyncPipe,
     ],
 })
-export class OptionsComponent extends UnSubscriber implements AfterViewInit {
+export class OptionsComponent implements AfterViewInit {
     @Input() public cardDocs!: IComDoc[];
     @Input() public cardExamples!: (TExample<IMaskOptions> | { _pipe: string })[];
 
     @ViewChildren('cards') public cards!: QueryList<ElementRef>;
 
-    public phone = '123456789';
-    public activeCardId = 1;
+    public readonly phone = '123456789';
     public readonly trackByPath = inject(TrackByService).trackBy('id');
+    public readonly activeCardId$: Observable<number> = inject(ScrollService).activeCard$;
+    public readonly openSourcePath = OpenSourcePath.OPEN_SOURCE;
 
     private readonly scrollService = inject(ScrollService);
-    private readonly router = inject(Router);
+    private readonly changeAccordionService = inject(ChangeAccordionService);
 
     public ngAfterViewInit(): void {
-        fromEvent(document, 'scroll')
-            .pipe(takeUntil(this.unsubscribe$$), debounceTime(100))
-            .subscribe(() => {
-                const scrollIdCard = this.cards.find((e) =>
-                    this.scrollService.isInViewport(e.nativeElement)
-                )?.nativeElement.id;
-                if (this.activeCardId !== Number(scrollIdCard)) {
-                    this.activeCardId = Number(scrollIdCard);
-                    this.router.navigate(['/'], {
-                        fragment: scrollIdCard,
-                    });
-                }
-            });
-        this.cards.changes.pipe(takeUntil(this.unsubscribe$$)).subscribe((elementRef) => {
-            this.router.navigate(['/'], {
-                fragment: '1',
-            });
-            const firstNativeElement: HTMLElement | null = document.getElementById(
-                elementRef.first.nativeElement.id
-            );
-            if (firstNativeElement) {
-                firstNativeElement.scrollIntoView({ behavior: 'smooth', block: 'end' });
-            }
-        });
+        this.scrollService.onScroll(this.cards);
+        this.changeAccordionService.onChangeAccordion(this.cards);
     }
 }
