+++ src/assets/images/open-source/separator.png
