+++ src/index.html
@@ -11,6 +11,6 @@
         <link rel="icon" type="image/x-icon" href="favicon.ico" />
     </head>
     <body>
-        <jsdaddy-open-source-root></jsdaddy-open-source-root>
+        <jsdaddy-open-source-root />
     </body>
 </html>
